setup with redo of current pattern with list-root
