+++ ui/lib/kv/addon/components/page/secret/details.hbs
@@ -1,9 +1,9 @@
 <KvPageHeader @breadcrumbs={{@breadcrumbs}} @model={{@model}} @pageTitle={{@model.path}}>
   <:tabLinks>
-    <LinkTo @route="secrets.secret.details" data-test-secrets-tab="Secret">Secret</LinkTo>
-    <LinkTo @route="secrets.secret.metadata.index" data-test-secrets-tab="Metadata">Metadata</LinkTo>
-    <LinkTo @route="secrets.secret.metadata.versions" data-test-secrets-tab="Metadata">Version History</LinkTo>
-    <LinkTo @route="secrets.secret.metadata.diff" data-test-secrets-tab="Metadata">Version Diff</LinkTo>
+    <LinkTo @route="secret.details" data-test-secrets-tab="Secret">Secret</LinkTo>
+    <LinkTo @route="secret.metadata.index" data-test-secrets-tab="Metadata">Metadata</LinkTo>
+    <LinkTo @route="secret.metadata.versions" data-test-secrets-tab="Metadata">Version History</LinkTo>
+    <LinkTo @route="secret.metadata.diff" data-test-secrets-tab="Metadata">Version Diff</LinkTo>
   </:tabLinks>
 
   <:toolbarFilters>
@@ -11,6 +11,6 @@
   </:toolbarFilters>
 
   <:toolbarActions>
-    <ToolbarLink @route="secrets.secret.edit" @type="add">Create new version</ToolbarLink>
+    <ToolbarLink @route="secret.edit" @type="add">Create new version</ToolbarLink>
   </:toolbarActions>
 </KvPageHeader>
