+++ ui/lib/kv/addon/components/page/secret/metadata-details.hbs
@@ -1,12 +1,12 @@
 <KvPageHeader @breadcrumbs={{@breadcrumbs}} @model={{@model}} @pageTitle={{@model.path}}>
   <:tabLinks>
-    <LinkTo @route="secrets.secret.details" data-test-secrets-tab="Secret">Secret</LinkTo>
-    <LinkTo @route="secrets.secret.metadata.index" data-test-secrets-tab="Metadata">Metadata</LinkTo>
-    <LinkTo @route="secrets.secret.metadata.versions" data-test-secrets-tab="Metadata">Version History</LinkTo>
-    <LinkTo @route="secrets.secret.metadata.diff" data-test-secrets-tab="Metadata">Version Diff</LinkTo>
+    <LinkTo @route="secret.details" data-test-secrets-tab="Secret">Secret</LinkTo>
+    <LinkTo @route="secret.metadata.index" data-test-secrets-tab="Metadata">Metadata</LinkTo>
+    <LinkTo @route="secret.metadata.versions" data-test-secrets-tab="Metadata">Version History</LinkTo>
+    <LinkTo @route="secret.metadata.diff" data-test-secrets-tab="Metadata">Version Diff</LinkTo>
   </:tabLinks>
 
   <:toolbarActions>
-    <ToolbarLink @route="secrets.secret.metadata.edit">Edit metadata</ToolbarLink>
+    <ToolbarLink @route="secret.metadata.edit">Edit metadata</ToolbarLink>
   </:toolbarActions>
 </KvPageHeader>
