+++ ui/lib/kv/addon/components/page/secret/version-diff.hbs
@@ -1,9 +1,9 @@
 <KvPageHeader @breadcrumbs={{@breadcrumbs}} @model={{@model}} @pageTitle={{@model.name}}>
   <:tabLinks>
-    <LinkTo @route="secrets.secret.details" data-test-secrets-tab="Secret">Secret</LinkTo>
-    <LinkTo @route="secrets.secret.metadata.index" data-test-secrets-tab="Metadata">Metadata</LinkTo>
-    <LinkTo @route="secrets.secret.metadata.versions" data-test-secrets-tab="Metadata">Version History</LinkTo>
-    <LinkTo @route="secrets.secret.metadata.diff" data-test-secrets-tab="Metadata">Version Diff</LinkTo>
+    <LinkTo @route="secret.details" data-test-secrets-tab="Secret">Secret</LinkTo>
+    <LinkTo @route="secret.metadata.index" data-test-secrets-tab="Metadata">Metadata</LinkTo>
+    <LinkTo @route="secret.metadata.versions" data-test-secrets-tab="Metadata">Version History</LinkTo>
+    <LinkTo @route="secret.metadata.diff" data-test-secrets-tab="Metadata">Version Diff</LinkTo>
   </:tabLinks>
 
 </KvPageHeader>
