+++ ui/lib/kv/addon/routes/list.js
@@ -4,17 +4,31 @@
  */
 
 import Route from '@ember/routing/route';
+import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { hash } from 'rsvp';
+import { normalizePath } from 'vault/utils/path-encoding-helpers';
 
-export default class KvSecretsRoute extends Route {
+export default class KvSecretsListRoute extends Route {
   @service store;
   @service secretMountPath;
 
+  nestedSecretName() {
+    const { name } = this.paramsFor('list');
+    return name ? normalizePath(name) : '';
+  }
+
+  // beforeModel() {
+  //   const secret = this.secretParam();
+  //   if (secret.endsWith('/')) {
+  //     this.router.replaceWith('secrets', secret + '/');
+  //   }
+  // }
   model() {
     // TODO add filtering and return model for query on kv/metadata.
+    const nestedSecret = this.nestedSecretName() || '';
     const backend = this.secretMountPath.currentPath;
-    const secrets = this.store.query('kv/metadata', { backend }).catch((err) => {
+    const secrets = this.store.query('kv/metadata', { backend, nestedSecret }).catch((err) => {
       if (err.httpStatus === 404) {
         return [];
       } else {
@@ -22,6 +36,7 @@ export default class KvSecretsRoute extends Route {
       }
     });
     return hash({
+      nestedSecret,
       secrets,
       backend,
     });
@@ -35,5 +50,11 @@ export default class KvSecretsRoute extends Route {
       { label: 'secrets', route: 'secrets', linkExternal: true },
       { label: resolvedModel.backend },
     ];
+  }
+  @action
+  willTransition() {
+    window.scrollTo(0, 0);
+    // ARG TODO not working
+    this.store.clearDataset('kv/metadata');
   }
 }
