+++ ui/lib/kv/addon/routes/secret/details.js
@@ -7,14 +7,14 @@ import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
 import { hash } from 'rsvp';
 
-export default class KvSecretMetadataIndexRoute extends Route {
+export default class KvSecretDetailsRoute extends Route {
   @service store;
   @service secretMountPath;
 
   model() {
-    // TODO return model for query on kv/metadata.
+    // TODO return model for query on kv/data.
     const backend = this.secretMountPath.get();
-    const { name } = this.paramsFor('secrets.secret');
+    const { name } = this.paramsFor('secret');
     return hash({
       path: name,
       backend,
@@ -25,10 +25,9 @@ export default class KvSecretMetadataIndexRoute extends Route {
     super.setupController(controller, resolvedModel);
 
     controller.breadcrumbs = [
-      { label: 'secrets', route: 'secrets', linkExternal: true },
-      { label: resolvedModel.backend, route: 'secrets' },
-      { label: resolvedModel.path, route: 'secrets.secret.details', model: resolvedModel.path },
-      { label: 'metadata' },
+      { label: 'secrets', route: 'lists', linkExternal: true },
+      { label: resolvedModel.backend, route: 'secret' },
+      { label: resolvedModel.path },
     ];
   }
 }
