+++ ui/lib/kv/addon/routes/secret/edit.js
@@ -9,7 +9,7 @@ export default class KvSecretEditRoute extends Route {
   model() {
     // TODO add model. Needs to include a queryParam for version.
     const backend = this.secretMountPath.get();
-    const { name } = this.paramsFor('secrets.secret');
+    const { name } = this.paramsFor('secret');
     return hash({
       path: name,
       backend,
@@ -20,9 +20,9 @@ export default class KvSecretEditRoute extends Route {
     super.setupController(controller, resolvedModel);
 
     controller.breadcrumbs = [
-      { label: 'secrets', route: 'secrets', linkExternal: true },
-      { label: resolvedModel.backend, route: 'secrets' },
-      { label: resolvedModel.path, route: 'secrets.secret.details', model: resolvedModel.path },
+      { label: 'secrets', route: 'list', linkExternal: true },
+      { label: resolvedModel.backend, route: 'secret' },
+      { label: resolvedModel.path, route: 'secret.details', model: resolvedModel.path },
       { label: 'edit' },
     ];
   }
