+++ ui/lib/kv/addon/routes/secret/metadata/index.js
@@ -7,26 +7,28 @@ import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
 import { hash } from 'rsvp';
 
-export default class KvSecretMetadataVersionsRoute extends Route {
+export default class KvSecretMetadataIndexRoute extends Route {
   @service store;
   @service secretMountPath;
 
   model() {
+    // TODO return model for query on kv/metadata.
     const backend = this.secretMountPath.get();
-    const parentModel = this.modelFor('secrets.secret.metadata');
+    const { name } = this.paramsFor('secret');
     return hash({
+      path: name,
       backend,
-      ...parentModel,
     });
   }
 
   setupController(controller, resolvedModel) {
     super.setupController(controller, resolvedModel);
+
     controller.breadcrumbs = [
-      { label: 'secrets', route: 'secrets', linkExternal: true },
-      { label: resolvedModel.backend, route: 'secrets' },
-      { label: resolvedModel.name, route: 'secrets.secret.details', model: resolvedModel.name },
-      { label: 'version history' },
+      { label: 'secrets', route: 'list', linkExternal: true },
+      { label: resolvedModel.backend, route: 'secret' },
+      { label: resolvedModel.path, route: 'secret.details', model: resolvedModel.path },
+      { label: 'metadata' },
     ];
   }
 }
