+++ ui/lib/kv/addon/templates/index.hbs
@@ -1,0 +1,2 @@
+{{page-title "Index"}}
+{{outlet}}
