+++ ui/lib/kv/addon/templates/list.hbs
@@ -1,0 +1,1 @@
+<Page::List @model={{this.model}} @breadcrumbs={{this.breadcrumbs}} @pageTitle={{this.pageTitle}} />
