+++ ui/lib/kv/addon/templates/secret/details.hbs
