+++ ui/lib/kv/addon/templates/secret/diff.hbs
@@ -1,0 +1,1 @@
+secret.diff
