+++ ui/lib/kv/addon/templates/secret/metadata/diff.hbs
