+++ ui/lib/kv/addon/templates/secret/metadata/edit.hbs
