+++ ui/lib/kv/addon/templates/secret/metadata/index.hbs
