+++ ui/lib/kv/addon/templates/secret/versions.hbs
@@ -1,0 +1,1 @@
+secret.versions
