--- ui/lib/kv/addon/templates/secrets/index.hbs
@@ -1,1 +1,0 @@
-<Page::Secrets @model={{this.model}} @breadcrumbs={{this.breadcrumbs}} @pageTitle={{this.pageTitle}} />
