--- ui/lib/kv/addon/templates/secrets/secret/diff.hbs
@@ -1,1 +1,0 @@
-secrets.secret.diff
