--- ui/lib/kv/addon/templates/secrets/secret/versions.hbs
@@ -1,1 +1,0 @@
-secrets.secret.versions
